# packages/api-spec/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Willet Cloud REST API",
    "version": "1.0.0",
    "description": "REST surface mirroring the Willet MCP tools. Authenticated with a Bearer API secret (wlt_...) minted in the dashboard. Reads use GET; writes use POST/PATCH/DELETE. Read-only secrets are limited to GET; project-scoped secrets may only address routes under their own project id."
  },
  "servers": [{ "url": "/api/v1" }],
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/organizations": {
      "get": {
        "operationId": "listOrganizations",
        "summary": "List organizations you belong to",
        "tags": ["organizations"],
        "parameters": [
          { "name": "include_inactive", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationList" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "operationId": "createOrganization",
        "summary": "Create an organization",
        "tags": ["organizations"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateOrgBody" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OrganizationBasic" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/organizations/{orgSlug}/projects": {
      "parameters": [{ "$ref": "#/components/parameters/OrgSlug" }],
      "get": {
        "operationId": "listProjects",
        "summary": "List projects in an organization",
        "tags": ["projects"],
        "responses": {
          "200": {
            "description": "Success",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectList" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "operationId": "createProject",
        "summary": "Create a project in an organization",
        "tags": ["projects"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectBody" } } }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectBasic" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API secret minted in the Willet dashboard, sent as `Authorization: Bearer wlt_...`."
      }
    },
    "parameters": {
...
      "OrgSlug": { "name": "orgSlug", "in": "path", "required": true, "description": "Organization slug.", "schema": { "type": "string" } },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request failed validation.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Missing, malformed, revoked, or expired API secret.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The secret lacks permission (read-only on a write, wrong scope, insufficient role).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The addressed resource does not exist.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
...
      "OrgRole": { "type": "string", "enum": ["owner", "admin", "member"] },
...
      "ProjectRole": { "type": "string", "enum": ["admin", "editor", "viewer"] },
      "OrgPlan": { "type": "string", "enum": ["free", "pro", "enterprise"] },
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" },
          "details": { "type": "array", "items": {} }
        }
      },
...
      "Organization": {
        "type": "object",
        "required": ["id", "name", "slug", "plan", "role", "createdAt", "deleteRequestedAt", "subscriptionActive"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "plan": { "$ref": "#/components/schemas/OrgPlan" },
          "role": { "$ref": "#/components/schemas/OrgRole" },
          "createdAt": { "type": "string" },
          "deleteRequestedAt": { "type": ["string", "null"] },
          "subscriptionActive": { "type": "boolean" }
        }
      },
      "OrganizationBasic": {
        "type": "object",
        "required": ["id", "name", "slug", "plan", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "plan": { "$ref": "#/components/schemas/OrgPlan" },
          "createdAt": { "type": "string" }
        }
      },
      "OrganizationList": {
        "type": "object",
        "required": ["count", "organizations"],
        "properties": {
          "count": { "type": "integer" },
          "organizations": { "type": "array", "items": { "$ref": "#/components/schemas/Organization" } }
        }
      },
...
      "Project": {
        "type": "object",
        "required": ["id", "name", "slug", "keyPrefix", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "keyPrefix": { "type": ["string", "null"] },
          "createdAt": { "type": "string" },
          "role": { "$ref": "#/components/schemas/ProjectRole" }
        }
      },
      "ProjectBasic": {
        "type": "object",
        "required": ["id", "name", "slug", "keyPrefix", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "keyPrefix": { "type": ["string", "null"] },
          "createdAt": { "type": "string" }
        }
      },
      "ProjectList": {
        "type": "object",
        "required": ["count", "projects"],
        "properties": {
          "count": { "type": "integer" },
          "projects": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } }
        }
      },
...
      "CreateOrgBody": {
        "type": "object",
        "required": ["name"],
        "properties": { "name": { "type": "string", "minLength": 1, "maxLength": 100 } }
      },
...
      "CreateProjectBody": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "key_prefix": { "type": "string", "minLength": 1, "maxLength": 50 }
        }
      },
...
    }
  }
}